backend/requirements.txt
django==2.2.19
python-dotenv==0.19.0
Pillow==8.3.1
PyJWT==2.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1
